backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
libsql-client==0.3.1
google-generativeai==0.8.3
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12
httpx==0.27.0
requests==2.31.0
protobuf==4.25.3
google.generativeai 
langchain 
langchain-community